Added a bit more customization.
Did some changes here.  The big one to sum it all up is that you don't
need to now set EMACS_ENV if you don't want to (but still should ;)).
The next one is that we set the current environment to a global variable
so that it can be checked in other libraries & packages.